+++ rust/vm/tests/cli.rs
@@ -4,5 +4,5 @@
 #[path = "cli"]
 mod cli {
     pub mod config_commands;
-    pub mod pkg_commands;
+    pub mod registry_commands;
 }